requirements.txt
Django==3.0.5
django-webpack-loader==0.6.0
pytz==2018.7

djangorestframework-jwt~=1.11.0
djangorestframework~=3.11.0